+++ contracts/interfaces/IGaugeController.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.19;
 
 interface IGaugeController {
-    function vote_for_gauge_weights(address gauge, uint weight) external;
+    function vote_for_gauge_weights(address gauge, uint256 weight) external;
 
     function gauge_types(address gauge) external view returns (int128);
 }
